Revert "Remove methods with `NotSupportedException`"
